refactor: 두 service 간 중복되는 코드를 DataSyncRepository로 이동
