+++ src/main/java/com/codeit/findex/data/AutoIndexDataSyncService.java
@@ -2,6 +2,7 @@ package com.codeit.findex.data;
 
 import com.codeit.findex.autosync.entity.AutoSyncConfig;
 import com.codeit.findex.autosync.repository.AutoSyncConfigRepository;
+import com.codeit.findex.autosync.service.AutoSyncConfigService;
 import com.codeit.findex.common.enums.SourceType;
 import com.codeit.findex.data.dto.Body;
 import com.codeit.findex.data.dto.Item;
@@ -12,6 +13,7 @@ import com.codeit.findex.indexinfo.repository.IndexInfoRepository;
 import com.codeit.findex.syncjob.entity.SyncJob;
 import com.codeit.findex.syncjob.repository.SyncJobRepository;
 import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -29,29 +31,18 @@ import java.nio.charset.StandardCharsets;
 import java.util.*;
 
 @Component
+@RequiredArgsConstructor
 public class AutoIndexDataSyncService {
-    @Value("${findex.api.key}")
-    private String apiKey;
 
-    @Value("${findex.api.url}")
-    private String apiUrl;
-
-    private final int MAX_RETRY = 3;
-
-    @Autowired
-    private IndexInfoRepository indexInfoRepository;
-
-    @Autowired
-    private IndexDataRepository indexDataRepository;
-
-    @Autowired
     private SyncJobRepository syncJobRepository;
 
-    @Autowired
     private AutoSyncConfigRepository autoSyncConfigRepository;
 
+    private final DataSyncRepository dataSyncRepository;
 
-    private final IndexApiParser indexApiParser = new IndexApiParser();
+    private final IndexApiParser indexApiParser;
+
+    private final int MAX_RETRY = 3;
 
     @Transactional
     public void fetchAndProcessAll() throws Exception {
@@ -80,9 +71,9 @@ public class AutoIndexDataSyncService {
                     pageNo = 1;
                     totalPages = Integer.MAX_VALUE;
                     while (pageNo <= totalPages) {
-                        StringBuilder urlBuilder = buildUrl(pageNo, numOfRows, trueIndexInfo.getIndexName(), beginBasDt);
+                        StringBuilder urlBuilder = dataSyncRepository.createUrl(pageNo, numOfRows, trueIndexInfo.getIndexName(), beginBasDt);
                         System.out.println("요청 url :" +  urlBuilder.toString());
-                        String responseJson = callApiWithRetry(urlBuilder);
+                        String responseJson = dataSyncRepository.callApiWithRetry(urlBuilder);
                         System.out.println("API 호출 시도: page=" + pageNo);
                         if(responseJson == null || responseJson.isEmpty()) {
                             System.out.println("responseJson에 문제 발생.");
@@ -100,7 +91,7 @@ public class AutoIndexDataSyncService {
                         List<Item> items = indexApiParser.parseItems(responseJson);
 
                         System.out.println("processItems 호출 - 아이템 수: " + items.size());
-                        storeItemsToDb(items);
+                        dataSyncRepository.storeItemsToDb(items);
 
                         pageNo++;
 
@@ -118,168 +109,6 @@ public class AutoIndexDataSyncService {
         }
 
     }
-
-    private String callApiWithRetry(StringBuilder urlBuilder) throws InterruptedException {
-        int retryCount = 0;
-
-        while (retryCount < MAX_RETRY) {
-            String responseJson = callApi(urlBuilder);
-
-            if(responseJson != null && !responseJson.isEmpty()) {
-                if(isValidResponse(responseJson)) {
-                    return responseJson;
-                } else {
-                    System.err.println("API 내부 오류 감지. 응답 내용 로그 출력:\n" + responseJson);
-                }
-            } else {
-                System.err.println("빈 응답 감지. 재시도 " + (retryCount+ 1));
-            }
-
-            retryCount++;
-            Thread.sleep(1000L * retryCount);
-
-        }
-
-        System.err.println("최대 재시도 도달. API 호출 실패.");
-        return null;
-    }
-
-    private StringBuilder buildUrl(int pageNo, int numOfRows, String idxNm, String beginBasDt) throws UnsupportedEncodingException {
-        StringBuilder urlBuilder = new StringBuilder(apiUrl);
-        urlBuilder.append("?serviceKey=").append(apiKey);
-        urlBuilder.append("&pageNo=").append(pageNo);
-        urlBuilder.append("&numOfRows=").append(numOfRows);
-        urlBuilder.append("&beginBasDt=").append(beginBasDt);
-        urlBuilder.append("&idxNm=").append(URLEncoder.encode(idxNm, "UTF-8"));
-        urlBuilder.append("&resultType=json");
-
-        return urlBuilder;
-    }
-
-    private String callApi(StringBuilder urlBuilder) {
-        try {
-            URL url = new URL(urlBuilder.toString());
-            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-            conn.setRequestMethod("GET");
-            conn.setRequestProperty("Content-type", "application/json");
-            int code = conn.getResponseCode();
-            System.out.println("Response code: " + code);
-            if(code != HttpURLConnection.HTTP_OK) {
-                throw new IOException("HTTP 오류 코드: " + code);
-            }
-            BufferedReader rd = new BufferedReader(new InputStreamReader(conn.getInputStream(), StandardCharsets.UTF_8));
-            StringBuilder sb = new StringBuilder();
-            String line;
-            while ((line = rd.readLine()) != null) {
-                sb.append(line);
-            }
-            rd.close();
-            conn.disconnect();
-
-            return sb.toString();
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private boolean isValidResponse(String responseJson) {
-        JSONObject json = new JSONObject(responseJson);
-        JSONObject header = json.getJSONObject("response").getJSONObject("header");
-        String resultCode = header.optString("resultCode", "");
-        String resultMsg = header.optString("resultMsg", "");
-
-        if(!resultCode.equals("00")) {
-            System.err.println("API 장애 또는 오류 반환:  코드 = " + resultCode + ", 메시지 = " +  resultMsg);
-            return false;
-        }
-        return true;
-    }
-
-    @Transactional
-    protected void storeItemsToDb(List<Item> items) {
-        Set<SyncJob> syncJobsList = new HashSet<>();
-        Set<IndexData> indexDataList = new HashSet<>();
-
-        for (Item item : items) {
-            IndexInfo indexInfo = null;
-            try {
-                indexInfo = indexInfoRepository.findByIndexName(item.getIdxNm())
-                        .orElseGet(() -> {
-                            IndexInfo newInfo = new IndexInfo();
-                            newInfo.setIndexClassification(item.getIdxCsf());
-                            newInfo.setIndexName(item.getIdxNm());
-                            newInfo.setSourceType(SourceType.OPEN_API);
-                            newInfo.setBasePointInTime(item.getBasPntm());
-                            newInfo.setBaseIndex(BigDecimal.valueOf(item.getBasIdx()));
-
-                            return newInfo;
-                        });
-                indexInfo.setEmployedItemsCount(item.getEpyItmsCnt());
-
-                indexInfo = indexInfoRepository.save(indexInfo);
-
-                IndexData indexData = indexDataRepository.findByIndexInfoAndBaseDate(indexInfo, item.getBasDt())
-                        .orElseGet(() -> {
-                            IndexData newData = new IndexData();
-                            newData.setSourceType(SourceType.OPEN_API);
-
-                            return newData;
-                        });
-                indexData.setIndexInfo(indexInfo);
-                indexData.setBaseDate(item.getBasDt());
-                indexData.setMarketPrice(item.getMkp());
-                indexData.setClosingPrice(item.getClpr());
-                indexData.setHighPrice(item.getHipr());
-                indexData.setLowPrice(item.getLopr());
-                indexData.setVersus(item.getVs());
-                indexData.setFluctuationRate(item.getFltRt());
-                indexData.setTradingQuantity(item.getTrqu());
-                indexData.setTradingPrice(item.getTrPrc());
-                indexData.setMarketTotalAmount(item.getLstgMrktTotAmt());
-
-                indexDataList.add(indexData);
-
-                SyncJob syncInfoJob = syncJobRepository.findByIndexInfoAndJobType(indexInfo, "INDEX_INFO")
-                    .orElseGet(() -> {
-                        SyncJob newJob = new SyncJob();
-                        newJob.setJobType("INDEX_INFO");
-                        newJob.setWorker("system");
-                        newJob.setResult("SUCCESS");
-
-                        return newJob;
-                    });
-                syncInfoJob.setIndexInfo(indexInfo);
-                syncInfoJob.setTargetDate(item.getBasDt());
-                syncJobsList.add(syncInfoJob);
-
-                SyncJob syncDataJob = syncJobRepository.findByIndexInfoAndJobTypeAndTargetDate(indexInfo, "INDEX_DATA", item.getBasDt())
-                        .orElseGet(() -> {
-                            SyncJob newJob = new SyncJob();
-                            newJob.setJobType("INDEX_DATA");
-                            newJob.setWorker("system");
-                            newJob.setResult("SUCCESS");
-
-                            return newJob;
-                        });
-                syncDataJob.setIndexInfo(indexInfo);
-                syncDataJob.setTargetDate(item.getBasDt());
-                syncJobsList.add(syncDataJob);
-            } catch (Exception e) {
-                Optional<SyncJob> syncInfoFailJob = syncJobRepository.findByIndexInfoAndJobType(indexInfo, "INDEX_INFO");
-                syncInfoFailJob.ifPresent(syncJob -> {syncJob.setResult("FAIL");});
-
-
-                Optional<SyncJob> syncDataFailJob = syncJobRepository.findByIndexInfoAndJobTypeAndTargetDate(indexInfo, "INDEX_DATA", item.getBasDt());
-                syncDataFailJob.ifPresent(syncJob -> {syncJob.setResult("FAIL");});
-
-                e.printStackTrace();
-            }
-        }
-
-        indexDataRepository.saveAll(indexDataList);
-        syncJobRepository.saveAll(syncJobsList);
-    }
-
 
 }
